backend/requirements.txt
alembic==1.0.2
amqp==2.5.2
aniso8601==7.0.0
arrow==0.13.1
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.6.3.0
blinker==1.4
celery==4.4.1
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.0
coverage==4.5.2
cryptography==2.7
dnspython==1.16.0
docopt==0.6.2
docxtpl==0.6.3
elasticsearch==7.0.0
entrypoints==0.3
flake8==3.7.8
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Celery==2.4.3
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
Flask-Mail==0.9.1
Flask-Migrate==2.3.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
flower==0.9.3
gunicorn==19.9.0
httpie==0.9.9
idna==2.7
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.6.8
lxml==4.5.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.1
Pillow==7.0.0
protobuf==3.6.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySnooper==0.3.0
python-dateutil==2.7.5
python-docx==0.8.7
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2019.1
redis==3.2.1
requests==2.20.0
rq==1.0
rq-dashboard==0.4.0
scipy==1.4.1
six==1.11.0
SQLAlchemy==1.2.13
tornado==5.1.1
urllib3==1.24
vine==1.3.0
Werkzeug==0.14.1
zipp==3.1.0
